Remove auto definition of normalized finder
Detection of the need for the normalized finder ultimately depends on refelction
into the table structure, which is a no-no in a class method.  Initialization, which
load the model that acts_as_trustee, fails if the table doesn't exist.  Migrations
are particularly affected by this failure and previously required a two-step process
of migrating and then making the model act_as_trustee.