Use standard Role#resource association
Replace three-level psuedo-association with a more standard one.  The only
value ever discerned for having the association in the first place was to allow
Role#to_s to adapt to it's resource's #to_s -but classes don't change names (!)
and the three-level association violates the principle of least surprise.

In addition, the semantic correlation between role and identity has been expunged.
Such a correlation, even if it exists in some client applications, is probably not
universal.